improved local collections
