+++ openeo/rest/connection.py
@@ -12,10 +12,6 @@ from pathlib import Path
 from typing import Dict, List, Tuple, Union, Callable, Optional, Any, Iterator
 from urllib.parse import urljoin
 
-import xarray as xr
-from glob import glob
-from pyproj import Transformer
-
 import requests
 from deprecated.sphinx import deprecated
 from requests import Response
@@ -28,6 +24,7 @@ from openeo.internal.graph_building import PGNode, as_flat_graph
 from openeo.internal.jupyter import VisualDict, VisualList
 from openeo.internal.processes.builder import ProcessBuilderBase
 from openeo.metadata import CollectionMetadata
+
 from openeo.rest import OpenEoClientException, OpenEoApiError, OpenEoRestError
 from openeo.rest.auth.auth import NullAuth, BearerAuth, BasicBearerAuth, OidcBearerAuth, OidcRefreshInfo
 from openeo.rest.auth.config import RefreshTokenStore, AuthConfig
@@ -43,9 +40,9 @@ from openeo.rest.service import Service
 from openeo.rest.udp import RESTUserDefinedProcess, Parameter
 from openeo.util import ensure_list, legacy_alias, dict_no_none, rfc3339, load_json_resource, LazyLoadCache, \
     ContextTimer, str_truncate
+from openeo.rest.localconnection import LocalConnection
 
 _log = logging.getLogger(__name__)
-
 
 def url_join(root_url: str, path: str):
     """Join a base url and sub path properly."""
@@ -207,202 +204,6 @@ class RestApiConnection:
 
     def __repr__(self):
         return "<{c} to {r!r} with {a}>".format(c=type(self).__name__, r=self._root_url, a=type(self.auth).__name__)
-
-    
-class LocalConnection():
-    """
-    Connection to no backend, for local processing.
-    """
-
-    def __init__(self,local_collections_path):
-        """
-        Constructor of LocalConnection.
-
-        :param local_collections_path: String path to the folder with the local collections in netCDF or geoTIFF
-        """
-        self.local_collections_path = local_collections_path.split('file://')[-1]
-        print(self.local_collections_path)
-        
-    def get_temporal_dimension(self,dims):
-        if 't' in dims:
-            return 't'
-        elif 'time' in dims:
-            return 'time'
-        elif 'temporal' in dims:
-            return 'temporal'
-        elif 'DATE' in dims:
-            return 'DATE'
-        else:
-            return None
-
-    def get_x_spatial_dimension(self,dims):
-        if 'x' in dims:
-            return 'x'
-        elif 'X' in dims:
-            return 'X'
-        elif 'lon' in dims:
-            return 'lon'
-        elif 'longitude' in dims:
-            return 'longitude'
-        else:
-            return None
-
-    def get_y_spatial_dimension(self,dims):
-        if 'y' in dims:
-            return 'y'
-        elif 'Y' in dims:
-            return 'Y'
-        elif 'lat' in dims:
-            return 'lat'
-        elif 'latitude' in dims:
-            return 'latitude'
-        else:
-            return None
-
-    def get_netcdf_metadata(self,file_path):
-        data = xr.open_dataset(file_path,chunks={})
-        t_dim = self.get_temporal_dimension(data.dims)
-        x_dim = self.get_x_spatial_dimension(data.dims)
-        y_dim = self.get_y_spatial_dimension(data.dims)
-
-        metadata = {}
-        metadata['stac_version'] = '1.0.0-rc.2'
-        metadata['type'] = 'Collection'
-        metadata['id'] = file_path
-        if 'title' in data.attrs:
-            metadata['title'] = data.attrs['title']
-        else:
-            metadata['title'] = file_path
-        if 'description' in data.attrs:
-            metadata['description'] = data.attrs['description']
-        else:
-            metadata['description'] = ''
-        if 'license' in data.attrs:
-            metadata['license'] = data.attrs['license']
-        else:
-            metadata['license'] = ''
-        providers = [{'name':'',
-                     'roles':['producer'],
-                     'url':''}]
-        if 'providers' in data.attrs:
-            providers[0]['name'] = data.attrs['providers']
-            metadata['providers'] = providers
-        elif 'institution' in data.attrs:
-            providers[0]['name'] = data.attrs['institution']
-            metadata['providers'] = providers
-        else:
-            metadata['providers'] = providers
-        if 'links' in data.attrs:
-            metadata['links'] = data.attrs['links']
-        else:
-            metadata['links'] = ''
-        x_min = data[x_dim].min().item(0)
-        x_max = data[x_dim].max().item(0)
-        y_min = data[y_dim].min().item(0)
-        y_max = data[y_dim].max().item(0)
-        
-        crs_present = False
-        bands = list(data.data_vars)
-        if 'crs' in bands:
-            bands.remove('crs')
-            crs_present = True
-            
-        if crs_present:
-            if 'crs_wkt' in data.crs.attrs:
-                transformer = Transformer.from_crs(data.crs.attrs['crs_wkt'], "epsg:4326")
-                lat_min,lon_min = transformer.transform(x_min,y_min)
-                lat_max,lon_max = transformer.transform(x_max,y_max)
-
-                t_min = str(data[t_dim].min().values)
-                t_max = str(data[t_dim].max().values)
-
-                extent = {'spatial': {'bbox': [[lon_min, lat_min, lon_max, lat_max]]},
-                         'temporal': {'interval': [[t_min,t_max]]}
-                 }
-        metadata['extent'] = extent
-        
-
-        
-        t_dimension = {t_dim: {'type': 'temporal', 'extent':[t_min,t_max]}}
-        x_dimension = {x_dim: {'type': 'spatial','axis':'x','extent':[x_min,x_max]}}
-        y_dimension = {y_dim: {'type': 'spatial','axis':'y','extent':[y_min,y_max]}}
-        if crs_present:
-            if 'crs_wkt' in data.crs.attrs:
-                x_dimension[x_dim]['reference_system'] = data.crs.attrs['crs_wkt']
-                y_dimension[y_dim]['reference_system'] = data.crs.attrs['crs_wkt']
-        
-        b_dimension = {}
-        if len(bands)>0:
-            b_dimension = {'bands': {'type': 'bands', 'values':bands}}
-        
-        metadata['cube:dimensions'] = {**t_dimension,**x_dimension,**y_dimension,**b_dimension}
-        
-        return metadata
-        
-    def get_netcdf_collections(self):
-        local_collections_netcdfs = glob(self.local_collections_path + '/*.nc')
-        local_collections_list = []
-        if len(local_collections_netcdfs)>0:
-            for local_netcdf in local_collections_netcdfs: 
-                netcdf_metadata = self.get_netcdf_metadata(local_netcdf)
-                local_collections_list.append(netcdf_metadata)
-        local_collections_dict = {'collections':local_collections_list}
-        return local_collections_dict
-        
-    def list_collections(self) -> List[dict]:
-        """
-        List basic metadata of all collections provided in the local collections folder.
-
-        .. caution::
-        :return: list of dictionaries with basic collection metadata.
-        """
-        data = self.get_netcdf_collections()["collections"]
-        return VisualList("collections", data=data)
-    
-    def describe_collection(self, collection_id: str) -> dict:
-        """
-        Get full collection metadata for given collection id.
-        
-        .. seealso::
-        
-            :py:meth:`~openeo.rest.connection.Connection.list_collection_ids`
-            to list all collection ids provided by the back-end.
-
-        :param collection_id: collection id
-        :return: collection metadata.
-        """
-        data = self.get_netcdf_metadata(collection_id)
-        return VisualDict("collection", data=data)
-    
-    def collection_metadata(self, name) -> CollectionMetadata:
-        # TODO: duplication with `Connection.describe_collection`: deprecate one or the other?
-        return CollectionMetadata(metadata=self.describe_collection(name))
-    
-    def load_collection(
-            self,
-            collection_id: str,
-            spatial_extent: Optional[Dict[str, float]] = None,
-            temporal_extent: Optional[List[Union[str, datetime.datetime, datetime.date]]] = None,
-            bands: Optional[List[str]] = None,
-            properties: Optional[Dict[str, Union[str, PGNode, Callable]]] = None,
-            fetch_metadata=True,
-    ) -> DataCube:
-        """
-        Load a DataCube by collection id.
-
-        :param collection_id: image collection identifier
-        :param spatial_extent: limit data to specified bounding box or polygons
-        :param temporal_extent: limit data to specified temporal interval
-        :param bands: only add the specified bands
-        :param properties: limit data by metadata property predicates
-        :return: a datacube containing the requested data
-        """
-        return DataCube.load_collection(
-            collection_id=collection_id, connection=self,
-            spatial_extent=spatial_extent, temporal_extent=temporal_extent, bands=bands, properties=properties,
-            fetch_metadata=fetch_metadata,
-        )
-
     
 class Connection(RestApiConnection):
     """
